pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airgradient"
version = "0.1.0"
description = "Python client library for AirGradient air quality monitors"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
mqtt = ["paho-mqtt>=1.6"]
dev = [
    "pytest>=7",
    "pytest-mock>=3",
    "responses>=0.23",
]

[project.scripts]
ag = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["airgradient*", "cli*"]
